Moved RecyclableViewWidget and NonRecyclableViewWidget to separate files
